feat: add the xen-hvm script to transform a PV domU into HVM domU
Add the xen-hvm scripts which takes an existing PV domU, as created by
xen-create and appy changes to transform it into an HVM domU.
The domU need to be booted to be transformed, so it must be valid and
functionnal.
All memory, vcpus and pinning settings are reset during the operation.